requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
plotly>=5.18.0
pyarrow>=14.0.0
numpy>=1.26.0
google-cloud-bigquery>=3.14.0
db-dtypes>=1.2.0
